+++ examples/sample/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @examples/sample
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @forma-js/react@1.2.1
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @forma-js/react@1.2.0
+
 ## 0.1.2
 
 ### Patch Changes